moving the new function to client manager and add comments on event recorder
